refactor(LikertScale): rename component
