+++ packages/components/src/LikertScale/LikertScale.module.scss
