+++ packages/components/src/LikertScale/_docs/LikertScale.mdx
@@ -1,10 +1,10 @@
 import { Canvas, Controls, Meta } from '@storybook/blocks'
 import { ResourceLinks, KAIOInstallation, RenameNotice } from '~storybook/components'
-import * as LikertScaleLegacyStories from './LikertScaleLegacy.stories'
+import * as LikertScale from './LikertScale.stories'
 
-<Meta of={LikertScaleLegacyStories} />
+<Meta of={LikertScale} />
 
-# LikertScaleLegacy
+# LikertScale
 
 <ResourceLinks
   sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/LikertScaleLegacy"
@@ -12,16 +12,14 @@ import * as LikertScaleLegacyStories from './LikertScaleLegacy.stories'
   designGuidelines="https://cultureamp.atlassian.net/wiki/spaces/DesignSystem/pages/3082060201/Likert+Scale"
 />
 
-<RenameNotice />
-
-<KAIOInstallation exportNames="LikertScaleLegacy" />
+<KAIOInstallation exportNames="LikertScale" />
 
 ## Overview
 
 Likert scale radio buttons let people select one option in a Likert scale ranging from strongly disagree to strongly agree.
 
-<Canvas of={LikertScaleLegacyStories.Playground} />
-<Controls of={LikertScaleLegacyStories.Playground} />
+<Canvas of={LikertScale.Playground} />
+<Controls of={LikertScale.Playground} />
 
 ## API
 
@@ -29,4 +27,4 @@ Likert scale radio buttons let people select one option in a Likert scale rangin
 
 Sets aria-required value on radiogroup for assistive technologies. An accessible label must be provided and validation must still be handled within implementations.
 
-<Canvas of={LikertScaleLegacyStories.IsRequired} />
+<Canvas of={LikertScale.IsRequired} />
