+++ packages/components/src/LikertScale/index.ts
@@ -1,0 +1,2 @@
+export * from './LikertScale'
+export * from './types'
