+++ packages/components/src/LikertScale/types.ts
