+++ packages/components/src/LikertScale/utils/determineSelectionFromKeyPress.ts
