--- packages/components/src/LikertScaleLegacy/index.ts
@@ -1,2 +1,0 @@
-export * from './LikertScaleLegacy'
-export * from './types'
