+++ packages/components/src/index.ts
@@ -31,7 +31,7 @@ export * from './KaizenProvider'
 export * from './Label'
 export * from './LabelledMessage'
 export * from './Link'
-export * from './LikertScaleLegacy'
+export * from './LikertScale'
 export * from './LinkButton'
 export * from './Loading'
 export * from './Menu'
